feat: refactor intos its own cli app
