+++ apps/firewall/src/credentials.ts
@@ -1,6 +1,6 @@
 import { readFileSync } from 'node:fs';
 
-/** Resolve Vercel project credentials from env (preferred) or the linked `.vercel/project.json`; throws if project/team/token are missing. Returns plain `string`s so callers see non-undefined types. Shared by the firewall scripts. */
+/** Resolve Vercel credentials for the project this tool manages (the SponsorSearch web app). Prefers env (VERCEL_PROJECT_ID/TEAM_ID/TOKEN from the repo-root .env.local); falls back to the web app's `vercel link` for the project/team ids. Throws if any are missing. Returns plain `string`s so callers see non-undefined types. */
 export function resolveVercelCredentials(): {
   projectId: string;
   teamId: string;
@@ -8,11 +8,15 @@ export function resolveVercelCredentials(): {
 } {
   let link: { projectId?: string; orgId?: string } = {};
   try {
+    // This tool manages the web app's WAF, so it reads the web app's Vercel link for the project/team ids.
     link = JSON.parse(
-      readFileSync(new URL('../.vercel/project.json', import.meta.url), 'utf8'),
+      readFileSync(
+        new URL('../../web/.vercel/project.json', import.meta.url),
+        'utf8',
+      ),
     );
   } catch {
-    // .vercel/project.json is gitignored / absent on fresh clones & CI — fall back to env.
+    // absent on fresh clones / CI — rely on the VERCEL_* env vars instead.
   }
   const projectId = process.env.VERCEL_PROJECT_ID ?? link.projectId;
   const teamId = process.env.VERCEL_TEAM_ID ?? link.orgId;
