+++ apps/web/package.json
@@ -69,7 +69,6 @@
     "@vitejs/plugin-react": "^6.0.2",
     "drizzle-kit": "^0.31.10",
     "h3": "^2.0.1-rc.20",
-    "ink": "^7.1.0",
     "playwright": "^1.60.0",
     "portless": "^0.13.0",
     "sharp": "^0.34.5",
