+++ package.json
@@ -19,8 +19,7 @@
     "db:migrate": "turbo run db:migrate --filter @ss/web",
     "db:push": "turbo run db:push --filter @ss/web",
     "db:studio": "turbo run db:studio --filter @ss/web",
-    "firewall:setup": "bun apps/web/scripts/setup-firewall.tsx",
-    "firewall:report": "bun apps/web/scripts/firewall-report.ts",
+    "firewall:setup": "bun apps/firewall/src/setup.tsx",
     "knip": "knip"
   },
   "dependencies": {},
